pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ants"
version = "0.1.0"
description = "Multi-agent collaboration: each ant is a container, AIP protocol, status API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.28",
    "docker>=7.1",
    "tenacity>=9.0",
    "pyyaml>=6.0",
    "openai>=1.55",
    "pymysql>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.8",
]
search = [
    "duckduckgo-search>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ants*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
